Sample Text File that Efective to Compress
Here is the sample that I mention in the description. My program could compress it from 48.300 bytes into 6.614 bytes or equal to 13.69% after being compressed. The reason why text become such good one to compress had been explained in the description file. If you can't find the file, it's in the master branch with filename "REAMDE.md". Thanks for reading my code :) I will be glad for your advice.